+++ cryptostore/data/influx.py
@@ -100,8 +100,8 @@ class InfluxDB(Store):
             r = requests.post(self.addr, data=c)
             # per influx docs, returns 204 on success
             if r.status_code != 204:
-                LOG.error("Influx: Failed to write data to %s - %d:%s", self.addr, r.status_code, r.reason)
-                raise EngineWriteError
+                LOG.warning("Influx: Failed to write data to %s - %d:%s", self.addr, r.status_code, r.reason)
+                raise Exception(r.status_code, r.reason)
         self.data = None
 
     def get_start_date(self, exchange: str, data_type: str, pair: str) -> float: